[refactor] 서비스 구현체 분리
